refactor: Include provider to make the form component smaller
